refactor: move clang-utils to lib folder
